requirements.txt
torch==2.7.1
torchvision==0.22.1
numpy==2.3.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
Pillow>=10
matplotlib>=3.6
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.16
scipy>=1.11
omegaconf>=2.3.0
requests>=2.28.1